+++ src/cldfgeojson/commands/glottolog_distance.py
@@ -1,6 +1,6 @@
 """
 Compute the distance between speaker areas in a CLDF dataset and corresponding Glottolog point
-coordinate.
+coordinates.
 
 Distances are given in "grid units", i.e. - since input is given as WGS 84 coordinates - in degrees
 Thus, near the equator, a distance of 1 would equal roughly 111km, while further away from the
@@ -9,7 +9,8 @@ equator it will be less.
 To get an overview of Glottolog distances for a dataset you may pipe the output to the csvstat
 tool:
 
-cldfbench geojson.glottolog_distance path/to/cldf --glottolog path/to/glottolog --format tsv | csvstat -t
+cldfbench geojson.glottolog_distance path/to/cldf --glottolog path/to/glottolog --format tsv | \
+csvstat -t
 
 You can also print the distances to the terminal using a tool like termgraph:
 
@@ -19,8 +20,9 @@ sed '/^$/d' | csvcut -t -c ID,Distance | csvsort -c Distance | csvformat -E | te
 from clldutils.clilib import Table, add_format
 from shapely.geometry import Point, shape, MultiPolygon
 from pycldf.cli_util import add_dataset, get_dataset, add_catalog_spec
-from pycldf.media import MediaTable
 from tqdm import tqdm
+
+from cldfgeojson.util import speaker_area_shapes
 
 
 def register(parser):
@@ -30,18 +32,8 @@ def register(parser):
 
 
 def run(args):
-    from cldfgeojson import MEDIA_TYPE
-
     ds = get_dataset(args)
-
-    geojsons = {}
-    for media in MediaTable(ds):
-        if media.mimetype == MEDIA_TYPE:
-            geojsons[media.id] = {
-                f['properties']['cldf:languageReference']: shape(f['geometry'])
-                for f in media.read_json()['features']
-                if f['properties'].get('cldf:languageReference')}
-
+    geojsons = speaker_area_shapes(ds, fix_geometry=True)
     gl_coords = {
         lg.id: Point(float(lg.longitude), float(lg.latitude))
         for lg in args.glottolog.api.languoids() if lg.longitude}
@@ -51,8 +43,10 @@ def run(args):
             if lg.cldf.glottocode in gl_coords:
                 if lg.cldf.speakerArea in geojsons:
                     shp = geojsons[lg.cldf.speakerArea][lg.cldf.id]
-                else:  # pragma: no cover
+                elif lg.cldf.speakerArea:  # pragma: no cover
                     shp = shape(lg.speaker_area_as_geojson_feature['geometry'])
+                else:
+                    continue
 
                 npolys = len(shp.geoms) if isinstance(shp, MultiPolygon) else 1
                 gl_coord = gl_coords[lg.cldf.glottocode]